pyproject.toml
[tool.poetry]
name = "img-to-video-svd"
version = "1.0.0"
description = "Image to video using stable video diffusion"
authors = ["Sumit Sahoo <sumitsahoo1988@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.0"
gradio = "^4.8.0"
rich = "^13.7.0"
setuptools = "^69.0.2"
replicate = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
